fix: create reusable cards for headers
