+++ screens/EditSignal/index.tsx
@@ -1,15 +1,12 @@
-import { StyleSheet, TouchableOpacity, View } from "react-native"
+import { StyleSheet, View } from "react-native"
 import Status from "@/components/cards/Status"
 import { CustomButton } from "@/components/ui/Button"
 import UserAvatar from "@/components/ui/UserAvatar"
-import CrossMark from "@/components/vectors/CrossMark"
 import { useNavigation } from "@react-navigation/native"
 import { NativeStackNavigationProp } from "@react-navigation/native-stack"
 import FriendsList from "@/components/lists/Friends"
 import Activity from "@/components/Activity"
 import { ScrollView } from "react-native-gesture-handler"
-import ShareCard from "@/components/Share"
-import CustomText from "@/components/ui/CustomText"
 import { theme } from "@/theme"
 import { useAuth } from "@/contexts/AuthContext"
 import { useStatus } from "@/contexts/StatusContext"
@@ -18,6 +15,10 @@ import { useEffect, useState } from "react"
 import { useMutation } from "@tanstack/react-query"
 import { useSignal } from "@/hooks/useSignal"
 import { StatusBar } from "expo-status-bar"
+import ConfirmAction from "@/components/ConfirmAction"
+import { onShare } from "@/utils/share"
+import ShareIcon from "@/components/vectors/ShareIcon"
+import Header from "@/components/cards/Header"
 
 type EditSignalScreenProps = NativeStackNavigationProp<
   RootStackParamList,
@@ -61,18 +62,7 @@ export default function EditSignal() {
   return (
     <View style={style.container}>
       <StatusBar style="dark" />
-      <View style={style.navBar}>
-        <CustomText style={style.headerText} fontWeight="bold">
-          Edit status
-        </CustomText>
-        <TouchableOpacity
-          onPress={() => {
-            navigation.goBack()
-          }}
-          style={style.CrossMarkButton}>
-          <CrossMark />
-        </TouchableOpacity>
-      </View>
+      <Header title=" Edit status" />
       <ScrollView
         keyboardShouldPersistTaps="always"
         contentContainerStyle={{
@@ -91,7 +81,13 @@ export default function EditSignal() {
           timeSlots={["NOW", "MORNING", "LUNCH", "AFTERNOON", "EVENING"]}
         />
         <FriendsList />
-        <ShareCard style={{ marginHorizontal: 20 }} />
+        <ConfirmAction
+          title=" Your friends are not on Wavv?"
+          description="Invite them to join you"
+          onPress={() => onShare(user?.username)}
+          icon={<ShareIcon />}
+          style={{ marginHorizontal: 20 }}
+        />
       </ScrollView>
       <CustomButton
         activeOpacity={0.8}
@@ -114,33 +110,11 @@ const style = StyleSheet.create({
     backgroundColor: theme.colors.white,
     position: "relative",
   },
-  navBar: {
-    justifyContent: "space-between",
-    alignItems: "center",
-    flexDirection: "row",
-    width: "100%",
-    paddingHorizontal: 20,
-    position: "absolute",
-    top: 44,
-    backgroundColor: theme.colors.white,
-    zIndex: 10,
-  },
-  headerText: {
-    flexGrow: 1,
-    textAlign: "center",
-    fontSize: 20,
-    lineHeight: 28,
-  },
   saveButton: {
     position: "absolute",
     bottom: 20,
     zIndex: 10,
     width: "90%",
     marginHorizontal: 20,
-  },
-  CrossMarkButton: {
-    padding: 5,
-    position: "absolute",
-    right: 20,
   },
 })
