+++ screens/Search/index.tsx
@@ -7,11 +7,8 @@ import {
 } from "react-native"
 import Input from "@/components/ui/Input"
 import UserInfo from "@/components/UserInfo"
-import ShareCard from "@/components/Share"
 import { CustomButton } from "@/components/ui/Button"
 import { useCallback, useEffect, useState } from "react"
-import CloseIcon from "@/components/vectors/CloseIcon"
-import { useNavigation } from "@react-navigation/native"
 import UserAvatar from "@/components/ui/UserAvatar"
 import CustomText from "@/components/ui/CustomText"
 import { User } from "@/types"
@@ -21,9 +18,13 @@ import api from "@/service"
 import { useFriends } from "@/hooks/useFriends"
 import { SafeAreaView } from "react-native-safe-area-context"
 import { StatusBar } from "expo-status-bar"
+import { onShare } from "@/utils/share"
+import ShareIcon from "@/components/vectors/ShareIcon"
+import ConfirmAction from "@/components/ConfirmAction"
+import { useAuth } from "@/contexts/AuthContext"
+import Header from "@/components/cards/Header"
 
 const FindFriends = () => {
-  const navigation = useNavigation()
   const [search, setSearch] = useState("")
   const [filteredUsers, setFilteredUsers] = useState<User[]>([])
   const [isLoadingFriend, setIsLoadingFriend] = useState<
@@ -32,6 +33,7 @@ const FindFriends = () => {
   const [allUsers, setAllUsers] = useState<User[]>([])
   const [isAnyFriendLoading, setIsAnyFriendLoading] = useState(false)
   const { fetchAllFriends } = useFriends()
+  const { user } = useAuth()
   const fetchUsers = useCallback(async () => {
     try {
       const response = await api.get(`/users`)
@@ -89,20 +91,10 @@ const FindFriends = () => {
     }
   }
 
-  const handleClose = () => navigation.goBack()
-
   return (
     <SafeAreaView style={styles.container}>
       <StatusBar style="dark" />
-      <View style={styles.header}>
-        <View style={styles.spacer} />
-        <CustomText size="lg" fontWeight="semibold">
-          Find Friends
-        </CustomText>
-        <TouchableOpacity onPress={handleClose}>
-          <CloseIcon color={theme.colors.black} />
-        </TouchableOpacity>
-      </View>
+      <Header title="Find Friends" />
       <Input
         variant="primary"
         textSize="base"
@@ -163,7 +155,12 @@ const FindFriends = () => {
         )}
 
         <View style={styles.share}>
-          <ShareCard />
+          <ConfirmAction
+            title=" Your friends are not on Wavv?"
+            description="Invite them to join you"
+            onPress={() => onShare(user?.username)}
+            icon={<ShareIcon />}
+          />
         </View>
       </ScrollView>
     </SafeAreaView>
@@ -176,18 +173,8 @@ const styles = StyleSheet.create({
     paddingTop: 20,
     backgroundColor: theme.colors.white,
   },
-  header: {
-    flexDirection: "row",
-    alignItems: "center",
-    justifyContent: "space-between",
-    paddingBottom: 22,
-    paddingHorizontal: 20,
-  },
   input: {
     marginHorizontal: 20,
-  },
-  spacer: {
-    width: 24,
   },
   friendsList: {
     marginTop: 10,
